+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Removed "optiboxes" command alias ('/skyboxify' is the only command now)
 - 1.21.2/3 & 1.21.4 now use the "custom_skybox" custom core shader for sky rendering
 - world4 folder is now classified as "aether:the_aether" dimension (Experimental)
+### Debug
+- "axis" is now pretty printed in the debug sky layer info screen
 - Added debug "dump" command which exports the active skyboxes into the "debug_skyboxify" folder
   - The data is the encoded value of the stored data via the CODEC
   - NOTE: If any value is the same as the default value, it won't be encoded